refactor(imports): update GitHub module imports across codebase
Update imports across the codebase to use the new modular GitHub submodules (types.ts, auth.ts, repositories.ts, commits.ts, utils.ts) instead of the original monolithic github.ts file. Fixed a duplicate function export issue with getInstallationManagementUrl.
