+++ TODO.md
@@ -220,7 +220,7 @@ This document outlines the detailed task breakdown for refactoring the GitPulse 
 
 **T025 - Update GitHub Imports**
 - Update imports across the codebase to use the new GitHub modules
-- Status: Not Started
+- Status: Completed [x]
 - Estimate: Medium
 - Depends On: T024
