+++ src/app/api/contributors/route.ts
@@ -1,7 +1,9 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getServerSession } from "next-auth";
 import { authOptions } from "../auth/[...nextauth]/route";
-import { fetchAllRepositories, fetchCommitsForRepositories, Commit } from "@/lib/github";
+import { Commit } from "@/lib/github/types";
+import { fetchAllRepositories } from "@/lib/github/repositories";
+import { fetchCommitsForRepositories } from "@/lib/github/commits";
 import { logger } from "@/lib/logger";
 import { generateETag, isCacheValid, notModifiedResponse, cachedJsonResponse, CacheTTL, generateCacheControl } from "@/lib/cache";
